+++ DigitalTwinLib/src/Model/Entities/Port.cpp
@@ -53,7 +53,7 @@ namespace DigitalTwin::Model {
         return resolveVariable(splittedAdress, 0);
     }
 
-    Variable* Port::resolveVariable(std::vector<std::string> domains, int index)
+    Variable* Port::resolveVariable(std::vector<std::string> domains, size_t index)
     {
         if (index >= domains.size())
             throw DigitalTwinAddressException();
